+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Add `TableQuery` hook for `InMemoryStorageProvider`.
+
 ## [0.1.0] - 2024-08-09
 
 ### Added